Support for Ubuntu 20.04
* Introduces the new method installing using subiquity and cloud-init,
* Increases the memory to 1GB for installer compatibility,
* Sets the SSH timeout and handshake limits to be high enough to be
  clear of the installer race condition.
* purge now breaks the VM disk, so this is gone,
* open-vm-tools is included for VMware, but we still need to install
  VirtualBox Additions

See: https://nickcharlton.net/posts/automating-ubuntu-2004-installs-with-packer.html

+++ scripts/vmtools.sh
@@ -4,6 +4,10 @@
 # Configure the relevant VM tools for this builder.
 ##
 
+distribution=$(lsb_release -si)
+version=$(lsb_release -sr)
+major_version=$(echo "$version" | awk -F . '{print $1}')
+
 case $PACKER_BUILDER_TYPE in
     'virtualbox-iso')
         echo "Installing VirtualBox Guest Additions..."
@@ -15,17 +19,23 @@ case $PACKER_BUILDER_TYPE in
         rm -f /home/vagrant/VBoxGuestAdditions.iso
     ;;
     'vmware-iso')
-        echo "Installing VMware Tools..."
-        apt-get -qy install fuse
-        mkdir -p /mnt/cdrom
-        mount -o loop /home/vagrant/linux.iso /mnt/cdrom
+				if [ "$distribution" = 'Ubuntu' ]; then
+					if [ "$major_version" -le "19" ]; then
+						echo "Installing VMware Tools..."
+						apt-get -qy install fuse
+						mkdir -p /mnt/cdrom
+						mount -o loop /home/vagrant/linux.iso /mnt/cdrom
 
-        cd /tmp
-        tar -zxpf /mnt/cdrom/VMwareTools-*.tar.gz -C /tmp/
-        /tmp/vmware-tools-distrib/vmware-install.pl --force-install --default
+						cd /tmp
+						tar -zxpf /mnt/cdrom/VMwareTools-*.tar.gz -C /tmp/
+						/tmp/vmware-tools-distrib/vmware-install.pl --force-install --default
 
-        umount /mnt/cdrom
-        rm -f /home/vagrant/linux.iso
+						umount /mnt/cdrom
+						rm -f /home/vagrant/linux.iso
+					else
+						echo "Skipping installing VMware Tools because open-vm-tools works..."
+					fi
+				fi
     ;;
     *)
         printf "Nothing to do for the %s builder type.\n" $PACKER_BUILDER_TYPE
